Flow type utils/ast
