+++ src/utils/editor/source-documents.js
@@ -9,7 +9,7 @@ import { getMode } from "../source";
 import type { Source } from "../../types";
 import { isWasm, getWasmLineNumberFormatter, renderWasmText } from "../wasm";
 import { resizeBreakpointGutter, resizeToggleButton } from "../ui";
-import type { SymbolDeclarations } from "../../workers/parser/getSymbols";
+import type { SymbolDeclarations } from "../../workers/parser";
 import type { SourceRecord } from "../../reducers/types";
 import SourceEditor from "./source-editor";
