+++ src/workers/parser/frameworks.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
-import getSymbols from "./getSymbols";
+import { getSymbols } from "./getSymbols";
 
 export function getFramework(sourceId) {
   const sourceSymbols = getSymbols(sourceId);
